move head tracker
